Move flutter/inspector/diagnostics to inspector/flutter/diagnostics to go with the other inspector code
