+++ packages/devtools_app/lib/src/inspector/flutter/diagnostics.dart
